refactor edgar_sample and add utility function
